fix(handler): handler parameters using more refined helper
When using the helper that takes a definition it would start evaluating
the constraint on the response within the definition even though it's
not needed for this helper. There might be a better way to get around
this, but this fix works and will help with the proof of concept.