feat(handler-aws): provide environment to compositions
This introduces the `environment` parameter to the type `LambdaHandlerCompositionFactory` which will be provided
the `process.env` by default. Additionally a utility method has
been added to the response of `factory()` which will allow for
the handler composition to be invoked with a custom environment.
This is useful for testing the handlers that require environment
variables as part of their service construction.